requirements.txt
unidecode
django-tables2
django-sitetree
